% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cellassignment_V2.R
\name{add_imagingcols_toSE}
\alias{add_imagingcols_toSE}
\title{Add your imaging columns to your SummarizedExperiment}
\usage{
add_imagingcols_toSE(
  SE,
  imaging_df,
  join_keys = c("Well", "Plate_ID"),
  columns_to_add = NULL
)
}
\arguments{
\item{SE}{Your SummarizedExperiment}

\item{imaging_df}{Your dataframe containing the imaging data}

\item{join_keys}{The columns you want to join by (match the wells by Well and Plate_ID)}

\item{columns_to_add}{Columns from your imaging dataset you're interested in adding to your SE}
}
\value{
An SE with your added imaging columns
}
\description{
Add your imaging columns to your SummarizedExperiment
}
