% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagingTools.R
\name{addImgPaths}
\alias{addImgPaths}
\title{Function that generates all the files paths for the images}
\usage{
addImgPaths(se, parent_folder, wellposition = 10, sep = "_")
}
\arguments{
\item{se}{SummarizedExperiment to where the folders should be stored}

\item{parent_folder}{Path to the TIF images outputed by Cluster Analysis, typically inside the folder Particle_Analysi}
}
\value{
updated se object
}
\description{
Function that generates all the files paths for the images
}
