% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colDatTools.R
\name{ag}
\alias{ag}
\title{Internal Aggregation function}
\usage{
ag(df, cols, fun = mean)
}
\arguments{
\item{df}{The data frame to be aggregated}

\item{cols}{by which columns to aggregate}
}
\value{
Aggregated df
}
\description{
Internal Aggregation function
}
