% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cellassignment_V2.R
\name{assign_cell_FINAL_valcol}
\alias{assign_cell_FINAL_valcol}
\title{Assign cell colors to your wells based on the imaging data using multiple value columns}
\usage{
assign_cell_FINAL_valcol(
  SE,
  channels = c("C1", "C2", "C3"),
  chan2analyze = c("C2", "C3"),
  join_keys = c("Well", "Plate_ID"),
  value_cols = c("Mean", "Area"),
  scale_suffix = "_scaled",
  color_output = "Color",
  classification_method = c("manual", "clustering"),
  threshold_mode = c("manual", "quantile"),
  red_thresh = 8.5,
  green_thresh = 5.2,
  c3_thresh = 5.6,
  red_quant = 0.5,
  green_quant = 0.75,
  c3_quant = 0.7,
  k_centers = 4,
  plot_umap = FALSE
)
}
\arguments{
\item{SE}{Your SummarizedExperiment}

\item{channels}{The different channels that your used in your experiment}

\item{chan2analyze}{The specific channels you're interested in using for the cell assignment (could also be all of them)}

\item{join_keys}{The columns you want to join by (match the wells by Well and Plate_ID)}

\item{scale_suffix}{The suffix that you want to have in your log1p scaled channel column names}

\item{color_output}{The name you wan to give the column that contains your cell color assignments}

\item{classification_method}{Set whether you prefer the cell assignment to be done based on clustering or based on manually-set thresholds}

\item{threshold_mode}{Should you choose manual classification, here you can set whether you want to set the threshold for a given quantile or a hard-cut threshold}

\item{red_thresh}{The manual hard threshold you want to set for the red channel}

\item{green_thresh}{The manual hard threshold you want to set for the green channel}

\item{c3_thresh}{The manual hard threshold you want to set for your third channel}

\item{red_quant}{The quantile-based threshold you want to set for the red channel}

\item{green_quant}{The quantile-based threshold you want to set for the green channel}

\item{c3_quant}{The quantile-based threshold you want to set for your third channel}

\item{k_centers}{The number of centers you want for your k-means clustering}

\item{plot_umap}{Whether you want the function to return the umap plot from the clustering if you ran clustering classification}

\item{value_col}{The features in your data that you're using to base the cell assignment on (could be Mean and Area for instance)}
}
\value{
A SE with your scaled channel columns and cell assignments
}
\description{
Assign cell colors to your wells based on the imaging data using multiple value columns
}
