% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagingTools.R
\name{brighten_image}
\alias{brighten_image}
\title{Helper function for imageval to brighten the images}
\usage{
brighten_image(img, factor = 1.5)
}
\arguments{
\item{factor}{Factor by which the image should be brightened}

\item{img.path}{Path to where your particle analysis images are located}
}
\value{
A brightened image
}
\description{
Helper function for imageval to brighten the images
}
