% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityTools.R
\name{checkAssay}
\alias{checkAssay}
\title{Add column-wise aggregation such as mean of any given assay and store it into colData}
\usage{
checkAssay(assayList, assayList.se)
}
\arguments{
\item{assayList}{list of assays in the SE}

\item{assayName}{list of assay names to check}
}
\value{
updated assayList
}
\description{
Add column-wise aggregation such as mean of any given assay and store it into colData
}
