% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagingTools.R
\name{image_paths}
\alias{image_paths}
\title{Function that generates all the files paths for the images}
\usage{
image_paths(parent_folder, idx, plate_ID, location)
}
\arguments{
\item{parent_folder}{Path to where your various experimental data is stored}

\item{idx}{The well(s) you want to look at}

\item{plate_ID}{The plate ID(s) in question}

\item{location}{The location of the plate_ID in the folder names (we assume here that it is always in the same spot)}
}
\value{
A brightened image
}
\description{
Function that generates all the files paths for the images
}
