% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagingTools.R
\name{imageval}
\alias{imageval}
\title{Plot the images with the different channels for a given well and plate}
\usage{
imageval(se, idx, plate_ID, green_slice = 2, red_slice = 3)
}
\arguments{
\item{idx}{Well number you want to look at, in the form of "H14" or "H09"}

\item{plate_ID}{Plate number you are interested in exploring}

\item{green_slice}{Identicate which slice is the green image (if working with GFP and ROMK for instance)}

\item{red_slice}{Identicate which slice is the red image}

\item{se_imagepath}{Path to your summarized experiment}
}
\value{
A plot of four images (brightfield, green and red channels and overlay of the two channels on top of the BF)
}
\description{
Plot the images with the different channels for a given well and plate
}
