% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cellassignment_V2.R
\name{plot_intensity_distribution}
\alias{plot_intensity_distribution}
\title{Generate density or ECDF plots based on your SE's log1p scaled channels to look at the general distribution}
\usage{
plot_intensity_distribution(
  SE,
  channels = c("C2", "C3"),
  value_col = "Mean",
  join_keys = c("Well", "Plate_ID"),
  plot_type = c("density", "ecdf"),
  channel_colors = c(C1 = "forestgreen", C2 = "lightcoral", C3 = "plum")
)
}
\arguments{
\item{SE}{Your SummarizedExperiment}

\item{channels}{The different channels that your used in your experiment}

\item{value_col}{The feature in your data that you're using to base the cell assignment on (could be Mean, Area, IntDen..)}

\item{join_keys}{The columns you want to join by (match the wells by Well and Plate_ID)}

\item{channel_colors}{The colors you want your channels to be in the plots}
}
\value{
A density or ECDF plot
}
\description{
Generate density or ECDF plots based on your SE's log1p scaled channels to look at the general distribution
}
