% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cellassignment_V2.R
\name{plot_well_assignment}
\alias{plot_well_assignment}
\title{Generate a well plate plot that shows the color assignment for the given wells}
\usage{
plot_well_assignment(
  SE,
  color_col = "Color_manual",
  title = NULL,
  colors = NULL
)
}
\arguments{
\item{SE}{Your SummarizedExperiment}

\item{color_col}{The column that contains your color assignment}

\item{title}{The title you want to give your plot}

\item{colors}{How you want the colors to be plotted}
}
\value{
A well plate plot showing the color assignment
}
\description{
Generate a well plate plot that shows the color assignment for the given wells
}
