% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cellassignment_V2.R
\name{plot_well_binary_assignment}
\alias{plot_well_binary_assignment}
\title{Generate a well plate plot that shows the whether the well is classified as induced, not-induced, ambiguous or unknown based on the well color assignment}
\usage{
plot_well_binary_assignment(
  SE,
  color_col = "Color_binary",
  values = c(red = "red", green = "green", navy = "navy", yellow = "yellow", orange =
    "orange", turquoise = "turquoise", white = "white", black = "grey"),
  title = NULL
)
}
\arguments{
\item{SE}{Your SummarizedExperiment}

\item{color_col}{The column that contains your color assignment}

\item{values}{The colors you want your categories to be assigned}

\item{title}{The title you want to give your plot}
}
\value{
A well plate plot showing the induction distribution
}
\description{
Generate a well plate plot that shows the whether the well is classified as induced, not-induced, ambiguous or unknown based on the well color assignment
}
