% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityTools.R
\name{plotAssayVSSweeps}
\alias{plotAssayVSSweeps}
\title{Wrapper function for plotting assays vs sweeps curves}
\usage{
plotAssayVSSweeps(
  se,
  assayList,
  rowCol,
  colorGroup = c(),
  wrapFormula = NULL,
  grouped = TRUE
)
}
\arguments{
\item{se}{SummarizedExperiment Object with reducedDim data}

\item{assayList}{assays to be plotted as y-value}

\item{rowCol}{numeric row Column to plot as x-value}

\item{colorGroup}{color grouping}

\item{wrapFormula}{formula to be given to facet_wrap}

\item{grouped}{boolean to have grouped value (mean of the grouping) or single wells}
}
\value{
a ggplot
}
\description{
Wrapper function for plotting assays vs sweeps curves
}
