% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSE.R
\name{prepareDF}
\alias{prepareDF}
\title{Prepare a DataControl Excel File into a DataFrame in long format
Make sure that it contains Online Analysis Results as well as QCs per sweep:
- Seal Resistance
- Series Resistance
- Capacitance
Make sure that the Excel file contains a "Nanion Chip Barcode" column}
\usage{
prepareDF(pathToDF)
}
\arguments{
\item{path}{Path to DataControl Excel}
}
\value{
A cleaned data.frame in long format
}
\description{
Prepare a DataControl Excel File into a DataFrame in long format
Make sure that it contains Online Analysis Results as well as QCs per sweep:
- Seal Resistance
- Series Resistance
- Capacitance
Make sure that the Excel file contains a "Nanion Chip Barcode" column
}
