% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se_hAG}
\alias{se_hAG}
\title{A SummarizedExperiment object containing merged Ephys and Imaging data
from a human adrenal gland dataset.
- 4 different measurements with 3 imaging datasets}
\format{
A SummarizedExperiment object with:
\describe{
  \item{assays}{The measurement matrix from electrophysiology}
  \item{colData}{Metadata on samples, including imaging features}
  \item{rowData}{Metadata on features/variables}
}
}
\source{
Generated from internal Excel and SQLite files
}
\usage{
se_hAG
}
\description{
A SummarizedExperiment object containing merged Ephys and Imaging data
from a human adrenal gland dataset.
- 4 different measurements with 3 imaging datasets
}
\keyword{datasets}
