% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{se_iN}
\alias{se_iN}
\title{A SummarizedExperiment object containing merged Ephys and Imaging data
from an iPSC-derived neuronal experiment
- 1 Ephys Experiment with 1 imaging dataset}
\format{
A SummarizedExperiment object with:
\describe{
  \item{assays}{The measurement matrix from electrophysiology}
  \item{colData}{Metadata on samples, including imaging features}
  \item{rowData}{Metadata on features/variables}
}
}
\source{
Generated from internal Excel and SQLite files
}
\usage{
se_iN
}
\description{
A SummarizedExperiment object containing merged Ephys and Imaging data
from an iPSC-derived neuronal experiment
- 1 Ephys Experiment with 1 imaging dataset
}
\keyword{datasets}
