% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccPreviewModule.R
\name{ccPreview}
\alias{ccPreview}
\alias{ccPreviewUI}
\alias{ccPreviewServer}
\title{CC Preview Shiny Module}
\usage{
ccPreviewUI(id)

ccPreviewServer(id)
}
\arguments{
\item{id}{Module namespace ID.}
}
\description{
Interactive AP detection preview embedded as a reusable Shiny module.
Can be embedded in any Shiny app (e.g. tinySEV).
}
