% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityTools.R
\name{colAG}
\alias{colAG}
\title{Add column-wise aggregation such as mean of any given assay and store it into colData}
\usage{
colAG(se, assayList, fun = mean, sweeps = row.names(se))
}
\arguments{
\item{se}{SummarizedExperiment Object with the Ephys-Data}

\item{fun}{to be changed, function used to aggregate, for now only mean available}

\item{sweeps}{which sweeps to take, tbd, default is all}

\item{assayNames}{list of assays to aggregate column-wise}
}
\value{
se with updated colData
}
\description{
Add column-wise aggregation such as mean of any given assay and store it into colData
}
