% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagingTools.R
\name{df_cleaned}
\alias{df_cleaned}
\title{Clean and normalize dataframe. Adds column and row identifiers and finds the
selections with the Hole-ROI}
\usage{
df_cleaned(df)
}
\arguments{
\item{df}{dataframe of image results}
}
\value{
A dataframe
}
\description{
Clean and normalize dataframe. Adds column and row identifiers and finds the
selections with the Hole-ROI
}
