% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityTools.R
\name{get_metric}
\alias{get_metric}
\title{Wrapper function to analyze IV-curves and extract Imax, Vmax and Vhalf.
Make sure that Vhalf is present in rowData}
\usage{
get_metric(se, assay = "Minima", inward = TRUE)
}
\arguments{
\item{se}{SummarizedExperiment Object with reducedDim data}

\item{assay}{assay to be analyzed}

\item{inward}{boolean stating whether the IV-curve shows in- or outward current}

\item{getErev}{boolean to try to get Erev or value where y = 0}
}
\value{
se with updated colData
}
\description{
Wrapper function to analyze IV-curves and extract Imax, Vmax and Vhalf.
Make sure that Vhalf is present in rowData
}
