% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityTools.R
\name{get_metric}
\alias{get_metric}
\title{Wrapper function to analyze IV-curves and extract Imax, Vmax and Vhalf}
\usage{
get_metric(df, well, parameter = "Minima", plot = FALSE)
}
\arguments{
\item{se}{SummarizedExperiment Object with reducedDim data}

\item{assayList}{assays that serves as the y-axis}

\item{rowCol}{numeric row Column that serves as x-axis}

\item{inward}{boolean stating whether the IV-curve shows in- or outward current}

\item{getErev}{boolean to try to get Erev or value where y = 0}
}
\value{
se with updated colData
}
\description{
Wrapper function to analyze IV-curves and extract Imax, Vmax and Vhalf
}
