% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colDatTools.R
\name{group_assignment}
\alias{group_assignment}
\title{Assign your wells to their corresponding groups (wildtype, hyperplasia, donors etc)}
\usage{
group_assignment(
  prepared_df,
  pattern = c("Conditions", "Alternating", "Block", "Manual", "Cycle"),
  manual_map = NULL,
  block_size = NULL,
  cycle_pattern = NULL
)
}
\arguments{
\item{prepared_df}{The data frame that you just filtered in filtered_df}

\item{pattern}{The pattern by which you separated your groups in the experiment}

\item{manual_map}{A named vector like c("A01" = "HP", "A02" = "WT", ...)}

\item{block_size}{Parameter that determines after how many rows the condition alternates from HP to WT}

\item{cycle_pattern}{A vector containing a custom, repeating sequence like the triplet c("HP", "WT", "HP")}
}
\value{
A data frame with group assignment
}
\description{
Assign your wells to their corresponding groups (wildtype, hyperplasia, donors etc)
}
