% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityTools.R
\name{plotDimRed}
\alias{plotDimRed}
\title{Wrapper function for plotting dimensionality plots
plot Dimensionality Reduction with ggplot}
\usage{
plotDimRed(se, redDim.method, colorColumns = character())
}
\arguments{
\item{se}{SummarizedExperiment Object with reducedDim data}

\item{redDim.method}{a single character parameter, either, "UMAP", "TSNE", "PCA"}

\item{colorColumns}{list of columns from colData adding colorings, every element with a different plot, default are clusters}
}
\value{
a ggplot
}
\description{
Wrapper function for plotting dimensionality plots
plot Dimensionality Reduction with ggplot
}
