% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colDatTools.R
\name{prep_df}
\alias{prep_df}
\title{Prepare the newly created data frame from the excel file: incorporation of the conditions, row and column columns etc}
\usage{
prep_df(wd_path, df)
}
\arguments{
\item{wd_path}{Path to the excel files location (including the meta data file)}

\item{df}{The data frame that you just created from the excel file you are looking at}
}
\value{
An adjusted data frame
}
\description{
Prepare the newly created data frame from the excel file: incorporation of the conditions, row and column columns etc
}
