% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colDatTools.R
\name{prep_excel}
\alias{prep_excel}
\title{Prepare the excel files from the Syncropatch Online Analysis}
\usage{
prep_excel(pathToExcel)
}
\arguments{
\item{pathToExcel}{Path to the excel file in question}
}
\value{
A data frame
}
\description{
Prepare the excel files from the Syncropatch Online Analysis
}
