% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSE.R
\name{prepareSE}
\alias{prepareSE}
\title{Prepare SummarizedExperiment Object from DataControl Excel-file}
\usage{
prepareSE(pathDF, conditionColumns = c("Compound"))
}
\arguments{
\item{pathDF}{Path to DataControl Excel files}

\item{conditionColumns}{array of columns that describe experimental conditions}
}
\value{
A SummarizedExperiment Object with OAs as assays
}
\description{
Prepare SummarizedExperiment Object from DataControl Excel-file
}
