% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityTools.R
\name{reducedDim.Cellwise}
\alias{reducedDim.Cellwise}
\title{Perform cell-wise dimensionality reduction based on assays and colData.
Results are stored in reducedDims}
\usage{
reducedDim.Cellwise(
  se,
  assayList = c(),
  colNames = c(),
  scaling = "within",
  byRow = FALSE,
  method = c("pca", "tsne", "umap"),
  k_clusters = 3
)
}
\arguments{
\item{se}{SummarizedExperiment Object with the Ephys-Data}

\item{colNames}{list of columns from colData}

\item{scaling}{option of which scaling to apply, within assay or all features together (scaling = "global"), default within assay}

\item{byRow}{scaling by row instead of by column (after transforming), default FALSE}

\item{method}{list of types of reductionality methods, default all (pca, tsne, umap)}

\item{k_clusters}{number of clusters}

\item{assayNames}{list of assays to include}
}
\value{
se with updated results
}
\description{
Perform cell-wise dimensionality reduction based on assays and colData.
Results are stored in reducedDims
}
