% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colDatTools.R
\name{row_aggregate_SE}
\alias{row_aggregate_SE}
\title{Create a SE with assays for the additions (Add1 - Add4)}
\usage{
row_aggregate_SE(
  se,
  metrics = c("Erev"),
  FUN = mean,
  na.rm = TRUE,
  periods = list(Add1 = 1:24, Add2 = 25:42, Add3 = 43:60, Add4 = 61:84),
  mode = c("by.period", "last.n"),
  last.n = 3
)
}
\arguments{
\item{se}{The summarized experiment you have created previously}

\item{FUN}{The method you would like to aggregate by}

\item{na.rm}{Remove NAs}

\item{periods}{Define your liquid periods}

\item{mode}{Aggregation mode by period or by last n sweeps}

\item{last.n}{The value of n if aggregating by the last n sweeps}

\item{assay_name}{The metric / features you want to have as assays}
}
\value{
A summarized experiment with the assays you set
}
\description{
Create a SE with assays for the additions (Add1 - Add4)
}
