% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colDatTools.R
\name{rowAG}
\alias{rowAG}
\title{Create a data frame with the feature you want to add to your summarized experiment's colData}
\usage{
rowAG(
  se,
  assay_name = "Erev",
  FUN = mean,
  na.rm = TRUE,
  periods = list(Add1 = 1:24, Add2 = 25:42, Add3 = 43:60, Add4 = 61:84),
  mode = c("by.period", "last.n"),
  last.n = 3
)
}
\arguments{
\item{se}{The summarized experiment you have created previously}

\item{assay_name}{The metric / features you want to add}

\item{FUN}{The method you would like to aggregate by}

\item{na.rm}{Remove NAs}

\item{periods}{Define your liquid periods}

\item{mode}{Aggregation mode by period or by last n sweeps}

\item{last.n}{The value of n if aggregating by the last n sweeps}
}
\value{
A data frame with the added feature in colData
}
\description{
Create a data frame with the feature you want to add to your summarized experiment's colData
}
